+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tables/validation/GameValidationTexts.java
@@ -159,11 +159,6 @@ public class GameValidationTexts {
         text = "The table contains an ALT Color file, but the DMD colorization is not enabled in the VPin Mame settings.";
         break;
       }
-      case CODE_ALT_COLOR_SERUM_INSTALLATION_FILES_MISSING: {
-        label = "Serum installation files missing.";
-        text = "The Serum ALT Color installation has missing files: " + state.getOptions().get(0);
-        break;
-      }
       case CODE_ALT_COLOR_DMDDEVICE_FILES_MISSING: {
         label = "DMD device files missing.";
         text = "Mandatory file not found to run ALT Color: " + state.getOptions().get(0);